sentinel-ai/requirements.txt
pandas==2.2.0
scikit-learn==1.4.0
psycopg2-binary==2.9.9
joblib==1.3.2
flask
flask-cors